Isolated Validation Functions, Fixed Frontend Error
Instead of having all the validation checks in the controllers, I created a
validation file and handled them there, making the code more seperated and
readable.

I also was forced to fix a error in CourseWidget on the frontend that was
taking the data sent from the backend and trying to compute with it even when
it was a JSON error, leading to the frontend to have an error.
 # Please enter the commit message for your changes. Lines starting