Created Util Folder, Will Need for Background Checking
Created the util folder to house the different utility we will need
to periodically check course status and notify the users. Created a
file for querying the Golang service on the status of class, a file
for sending the email to the user, and a file for handling the
subscription queries.